+++ src/Element/Element.php
@@ -53,6 +53,7 @@ abstract class Element extends Component implements ElementInterface
     use Concerns\HasCanonical;
     use Concerns\HasControlPanelUI;
     use Concerns\HasCustomFields;
+    use Concerns\HasDeletionBlockers;
     use Concerns\HasGqlType;
     use Concerns\HasLifecycleHooks;
     use Concerns\HasPreviewTargets;
@@ -231,7 +232,7 @@ abstract class Element extends Component implements ElementInterface
      */
     private ?array $_attributeNames = null;
 
-    private bool $_initialized = false;
+    private bool $_trackDirtyFields = false;
 
     /**
      * @see toArray()
@@ -255,7 +256,7 @@ abstract class Element extends Component implements ElementInterface
             $this->_savedTitle = $this->title;
         }
 
-        $this->_initialized = true;
+        $this->_trackDirtyFields = true;
     }
 
     public function __clone()
@@ -301,12 +302,19 @@ abstract class Element extends Component implements ElementInterface
         if (str_starts_with($name, 'field:')) {
             return app(Fields::class)->isKnownFieldHandle(substr($name, 6));
         }
+
         if ($name === 'title') {
             return true;
         }
+
+        if (isset($this->_generatedFieldValues[$name])) {
+            return true;
+        }
+
         if ($this->hasEagerLoadedElements($name)) {
             return true;
         }
+
         if (parent::__isset($name)) {
             return true;
         }